Prepare release 0.6.6
This release has no functional changes. It is intended to be the
first release where upload of the traceability report works.

It is not great to make a new release simply to trigger the upload,
but I would like to avoid re-building a new git revision
(d28705a5080cb189c3645d815fcaaacf54bd04b3) with the existing 0.6.5 tag.
This would mean we have a traceability report for a build which does not
exactly match the revision uploaded to pypi.org, even if there no
functional changes.