Travis: add pypi upload as separate stage
The previous changes moved the pypi upload into the system test
stage. That stage is using Windows, which causes the pypi upload to fail.

With this commit, the traceability report is now still uploaded from the
system test stage while the pypi upload happens in its own stage.

This commit also removes `skip_existing: true` and `skip_cleanup: true` from
the pypi deploy step. The former is removed because it should not happen that
an existing upload is performed twice. If a build has a problem and the upload
still succeeded and we want to re-upload, we should do that with a fresh
version number.

The latter is removed because the stage will obtain a fresh checkout.