Travis: Explicitly set TOXENV
We used `tox-travis` to call the correct tox environment
(with the correct Python environment) on Travis. However,
the macOS and Windows builds did not use the `language: python`
setting on Travis. For these builds, tox was called but no
tests were executed.

This commit explicitly sets TOXENV for all builds in the test
stage to ensure that all tests run.