Add support for custom Job and Deployment IDs
This commit adds support for custom Job and Deployment IDs to be
passed to the Data Attribute Recommendation service when creating
the respective Job or Deployments.

This change is not generally available in the service, but the SDK
supports it already.

Additionally, this PR updates all dependencies and checks to latest
along with any fixes required by these updates.