Properly specify compatible version of dependencies
Commit 855c090cd9bbeeb345152ab6d535fbd85e444b61 in release 0.8.0 introduced
use of the `~=` operator to declare package dependencies. Sadly, this introduced
a bug where now the package depended on `requests >= 2.20.0; requests < 2.21.0` due
to the way the "compatible release operator" is specified in PEP-440.

The notion of "compatible release" in PEP-440 thus does not strictly follow semantic
versioning.

This commit updates the dependencies to once again allow updates to any minor
version.