Raise DARHTTPException also for retried status codes
For retried HTTP status codes such as 500, the `requests` library would
internally raise a `RetryError`. As a consequence, the `DARSession`
did not raise the `DARHTTPException` which contains important
debugging information.

This commit fixes issue #104 by configuring `requests` (rather, `urllib3`)
to return the last response instead of raising an exception. The response
contains the HTTP 500 status code which is then converted to a `DARHTTPException`
in the `DARSession`.

The initial approach to handle this issue was to catch the `RetryException` and
extract the last `Response` object from the exception. However, this is not possible
due to a limitation in `urllib3`: https://github.com/urllib3/urllib3/issues/2096

Fixes #104